Separate CTAP2 ceremonies from client request preparation
